Added Update Functionality
Each card has unique ID provided by firebase Real-time database. >
When edit item is clicked , pass ID as props to the modal component. >modal on useEffect get record from DB with the ID, returns and populates fields of current sheet. >
click update , updates the value. on the DB. >
component listening to the DB listens for the change and re-renders the component and loads records again.